% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{BhattacharyyaDist}
\alias{BhattacharyyaDist}
\title{Bhattacharyya distance}
\usage{
BhattacharyyaDist(x, adj = 0.01)
}
\arguments{
\item{x}{Matrix}

\item{adj}{Small quantity added to avoid indefinite log(0) values. DEFAULT=0.001}
}
\value{
Distance R object
}
\description{
Bhattacharyya distance core function
}
