% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BhattacharyyaDist_rcpp}
\alias{BhattacharyyaDist_rcpp}
\title{Bhattacharyya Distance via Rcpp}
\arguments{
\item{x}{n X p character matrix.}

\item{offset}{small offset for log(0*0) cases.}
}
\value{
Distance matrix between rows.
}
\description{
Computes pairwise Bhattacharyya distance between rows.
}
\details{
Bhattacharyya Distance
}
