% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{BhattacharyyaDistRcpp}
\alias{BhattacharyyaDistRcpp}
\title{Bhattacharyya distance via Rcpp}
\usage{
BhattacharyyaDistRcpp(X, offset = 1e-08)
}
\arguments{
\item{X}{character matrix of categorical data}

\item{offset}{small value to avoid cases when log(0*0)}
}
\value{
A numeric matrix of pairwise distances
}
\description{
Bhattacharyya distance core function
}
