% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ChisqDist_rcpp}
\alias{ChisqDist_rcpp}
\title{Chi-square Distance via Rcpp}
\arguments{
\item{x}{n X p character matrix.}
}
\value{
Distance matrix between rows.
}
\description{
Computes pairwise Chi-square distance between rows.
}
\details{
Chi-square Distance
}
