% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{ChisqDistRcpp}
\alias{ChisqDistRcpp}
\title{Chi-square distance via Rcpp}
\usage{
ChisqDistRcpp(X)
}
\arguments{
\item{X}{character matrix of categorical data}
}
\value{
A numeric matrix of pairwise distances
}
\description{
Chi-square distance core function
}
