% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteClusGapFuns.R
\name{clusGapDiscr}
\alias{clusGapDiscr}
\title{Discrete application of clusGap}
\usage{
clusGapDiscr(
  x,
  clusterFUN,
  K.max,
  B = nrow(x),
  value.range = "DS",
  verbose = interactive(),
  distName = "hamming",
  useLog = TRUE,
  dataClass = "nom",
  ...
)
}
\arguments{
\item{x}{A matrix object specifying category attributes in the columns and observations in the rows.}

\item{clusterFUN}{Character string with one of the available clustering implementations.
Available options are: 'pam' (default) from \code{cluster::pam}, 'diana' from \code{cluster::diana}, 'fanny' from \code{cluster::fanny},
'agnes-\{average, single, complete, ward, weighted\}' from \code{cluster::fanny},
'hclust-\{ward.D, ward.D2, single, complete, average, mcquitty, median, centroid\}' from \code{stats::hclust},
'kmodes' from \code{klar::kmodes} (\code{iter.max = 10}, \code{weighted = FALSE} and \code{fast= TRUE}).
'kmodes-N' enables to run the \code{kmodes} algorithm with a given number N of iterations where \code{iter.max = N}.}

\item{K.max}{Integer. Maximum number of clusters \code{k} to consider}

\item{B}{Number of bootstrap samples. By default B = nrow(x).}

\item{value.range}{String character vector or a list of character vector with the length matching the number of columns (nQ) of the array.
A vector with all categories to consider when bootstrapping the null distribution sample (KS: Known Support option).
By DEFAULT vals=NULL, meaning unique range of categories found in the data will be used when drawing the null (DS: Data Support option).
If a character vector of categories is provided, these values would be used for the null distribution drawing across the array.
If a list with category character vectors is provided, it has to have the same number of columns as the input array. The order of list element corresponds to the array's columns.}

\item{verbose}{Integer or logical. Determines whether progress output should printed while running. By DEFAULT one bit is printed per bootstrap sample.}

\item{distName}{String. Name of categorical distance to apply.
Available distances: 'bhattacharyya', 'chisquare', 'cramerV', 'hamming' and 'hellinger'.}

\item{useLog}{Logical. Use log function after estimating \code{W.k}. Following the original formulation \code{useLog=TRUE} by default.}

\item{dataClass}{character. Either 'nom' for nominal or 'ord' for ordinal.}

\item{...}{optionally further arguments for \code{FUNcluster()}}
}
\value{
a matrix with K.max rows and 4 columns, named "logW", "E.logW", "gap", and "SE.sim",
where gap = E.logW - logW, and SE.sim correspond to the standard error of \code{gap}.
}
\description{
Based on the implementation of the function found in the \code{cluster} R package.
}
