% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteClusGapFuns.R
\name{clusGapDiscr0}
\alias{clusGapDiscr0}
\title{Discrete application of clusGap - core function.}
\usage{
clusGapDiscr0(
  x,
  FUNcluster,
  K.max,
  B = nrow(x),
  value.range = "DS",
  verbose = interactive(),
  distName = "hamming",
  useLog = TRUE,
  Input2Alg = "distMatr",
  dataClass = "nom",
  offset = 0,
  ...
)
}
\arguments{
\item{x}{A matrix object specifying category attributes in the columns and observations in the rows.}

\item{FUNcluster}{a function that accepts as first argument a matrix like \code{x}; second argument specifies number of \code{k} (k=>2) clusters
This function should return a list with a component named \code{cluster}, a vector of length \code{n=nrow(x)} of integers from \code{1:k} indicating observation cluster assignment.
Make sure \code{FUNcluster} and \code{Input2Alg} agree.}

\item{K.max}{Integer. Maximum number of clusters \code{k} to consider}

\item{B}{Number of bootstrap samples. By default \code{B = nrow(x)}.}

\item{value.range}{A length 1 character string, a character string vector or a list of character vector with the length matching the number of columns (nQ) of the array.
By DEFAULT value.range = 'DS' (Data Support null model).
A vector with all categories (either character for nominal or integer ordinal data) to consider when bootstrapping the null distribution sample (KS: Known Support option).
If a list with category vectors is provided, it has to have the same number of columns as the input array. The order of list element corresponds to the array's columns.}

\item{verbose}{Integer or logical. Determines whether progress output should printed while running. By DEFAULT one bit is printed per bootstrap sample.}

\item{distName}{String. Name of categorical distance to apply.
Available distances: 'bhattacharyya', 'chisquare', 'cramerV', 'hamming' and 'hellinger'.}

\item{useLog}{Logical. Use log function after estimating \code{W.k}. Following the original formulation \code{useLog=TRUE} by default.}

\item{Input2Alg}{Specifies the kind of input provided to the algorithm function in \code{FUNcluster}.
For algorithms that only accept a distance matrix use \code{'distMatr'} option (default).
For algorithms that require the dataset and a prespecified distance function (e.g. \code{stats::dist}) use the \code{'distFun'} option.
This case the distance function is defined internally and determined by parameter \code{distName}.}

\item{dataClass}{character. Either 'nom' for nominal or 'ord' for ordinal.}

\item{offset}{numerical. A small constant value added to W.k to avoid NAs when running \code{useLog=TRUE}
for clusters with extremely low variability. \code{offset = 0} is set by default.}

\item{...}{optionally further arguments for \code{FUNcluster()}}
}
\value{
a matrix with K.max rows and 4 columns, named "logW", "E.logW", "gap", and "SE.sim",
where gap = E.logW - logW, and SE.sim correspond to the standard error of \code{gap}.
}
\description{
Based on the implementation of the function found in the \code{cluster} R package.
This function assumes that all attributes have identical categories.
}
