% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFuns.R
\name{clusGapDiscrHeat}
\alias{clusGapDiscrHeat}
\title{Resulting clustergram/heatmap.
Function to display a categorical data matrix given a user defined number of clusters `nCl`, a categorical distance `distName` and a predefined clustering method `FUNcluster`.
The output displays a heatmap separating and color-labelling resulting clusters vertically in the rows and allowing unsupervised clustering on questions in the columns. Each cell is colored according to the categorical values provided or found in the data.
The clustergram is based on the `pheatmap` function from the pheatmap R package. Thus, any parameter found in pheatmap can be specified to `clusGapDiscrHeat`.
This function can be used to examine number of clusters before running `clusGapDiscrHeat` but also after number of clusters is determined.}
\usage{
clusGapDiscrHeat(x, FUNcluster, nCl, distName, catVals = NULL, ...)
}
\arguments{
\item{x}{matrix object or data.frame}

\item{FUNcluster}{a function that accepts as first argument a matrix like `x`; second argument specifies number of `k` (k=>2) clusters
This function returns a list with a component named `cluster`, a vector of length `n=nrow(x)` of integers from `1:k` indicating observation cluster assignment.}

\item{nCl}{number of clusters to plot}

\item{distName}{Name of categorical distance to apply.
Available distances: 'bhattacharyya', 'chisquare', 'cramerV', 'hamming' and 'hellinger'.}

\item{catVals}{All categorical values to consider for the plot. If `NULL` the function will extract all unique categorical values in `x`.}
}
\value{
pheatmap output.
}
\description{
Resulting clustergram/heatmap.
Function to display a categorical data matrix given a user defined number of clusters `nCl`, a categorical distance `distName` and a predefined clustering method `FUNcluster`.
The output displays a heatmap separating and color-labelling resulting clusters vertically in the rows and allowing unsupervised clustering on questions in the columns. Each cell is colored according to the categorical values provided or found in the data.
The clustergram is based on the `pheatmap` function from the pheatmap R package. Thus, any parameter found in pheatmap can be specified to `clusGapDiscrHeat`.
This function can be used to examine number of clusters before running `clusGapDiscrHeat` but also after number of clusters is determined.
}
