% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFuns.R
\name{clusterFunSel}
\alias{clusterFunSel}
\title{Clustering generating function}
\usage{
clusterFunSel(clustFun)
}
\arguments{
\item{clustFun}{A character string with the following possible options:
'pam' (default) from \code{cluster::pam}, 'diana' from \code{cluster::diana}, 'fanny' from \code{cluster::fanny},
'agnes-\{average, single, complete, ward, weighted\}' from \code{cluster::agnes},
'hclust-\{ward.D, ward.D2, single, complete, average, mcquitty, median, centroid\}' from \code{base::hclust},
'kmodes' from \code{klar::kmodes} (\code{iter.max = 10}, \code{weighted = FALSE} and \code{fast= TRUE}).
'kmodes-N' enables to run the \code{kmodes} algorithm with a given number N of iterations where \code{iter.max = N}.}
}
\value{
An object of class kmodes as found in \code{klaR} packages.
An additional component specifies the categorical distance function found in \code{distFun}.
}
\description{
A function that generates formatted algorithmic functions that can be plugged
to enable run a wide variety of clustering algorithm for \code{clusGapDiscr} function.
}
