% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concussion}
\alias{concussion}
\title{Concussion Data}
\format{
## `data.frame`

\describe{
  \item{Q1: Headache}{Headache}
  \item{Q2: Nausea}{Nausea}
  \item{Q3: Balance problems}{Balance problems}
  \item{Q4: Dizziness}{Dizziness}
  \item{Q5: Fatigue}{Fatigue}
  \item{Q6: Sleep more}{Sleeping more than usual}
  \item{Q7: Drowsiness}{Drowsiness}
  \item{Q8: Sensibility to light}{Sensibility to light}
  \item{Q9: Sensibility to noice}{Sensibility to noice}
  \item{Q10: Irritability}{Irritability}
  \item{Q11: Sadness}{Sadness}
  \item{Q12: Nervousness}{Nervousness/Anxiousness}
  \item{Q13: More emotional}{Feeling more emotional}
  \item{Q14: Feeling slowed down}{Feeling slowed down}
  \item{Q15: Feeling mentally foggy}{Feeling mentally foggy}
  \item{Q16: Difficulty concentrating}{Difficulty concentrating}
  \item{Q17: Difficulty remembering}{Difficulty remembering}
  \item{Q18: Visual problem}{Visual problems}
  \item{Q19: Confusion}{Confusion}
  \item{Q20: Feeling clumsy}{Feeling clumsy}
  \item{Q21: Answer slowlier}{Answer slowlier}
}
}
\usage{
concussion
}
\description{
A data frame with 109 observations and 21 questions.
Severity rating recorded as categorical responses from c1 (none) to c7 (severe).
}
\keyword{datasets}
