% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{cramersVmod}
\alias{cramersVmod}
\title{Cramer's V modified pairwise vector function based on the function found in lsr package}
\usage{
cramersVmod(x, y)
}
\arguments{
\item{x}{vector of size n}

\item{y}{vector of size n}
}
\value{
numerical value
}
\description{
This is simple wrapper of the usual chisq.test function.
This is actually an adjusted version of the pi = sqrt(Chisq2/N)
guaranteeing that values are within 0 (no association) and 1 (association)
}
