% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{dissabs}
\alias{dissabs}
\title{####################
Ordinal distances ##
####################
Absolute distance}
\usage{
dissabs(X)
}
\arguments{
\item{X}{Numerical matrix with integer/ordinal values.}
}
\value{
Distance matrix
}
\description{
Absolute distance for ordinal data
}
