% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{dissbhattacharyya}
\alias{dissbhattacharyya}
\title{Bhattacharyya's distance (wrapper)}
\usage{
dissbhattacharyya(X)
}
\arguments{
\item{X}{Matrix}
}
\value{
Distance R object
}
\description{
Wrapper of `BhattacharyyaDist`
}
