% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{disschisquare}
\alias{disschisquare}
\title{Chi-square distance wrapper function}
\usage{
disschisquare(X, na.rm = TRUE)
}
\arguments{
\item{X}{Matrix}

\item{na.rm}{logical}
}
\value{
Distance R object
}
\description{
Chi-square distance wrapper function
}
