% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{disscramerv}
\alias{disscramerv}
\title{Cramer's V distance wrapper function}
\usage{
disscramerv(X, na.rm = TRUE)
}
\arguments{
\item{X}{Matrix}

\item{na.rm}{logical}
}
\value{
Distance R object
}
\description{
Cramer's V distance wrapper function
}
