% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{dissgordon}
\alias{dissgordon}
\title{####################
Ordinal distances ##
####################
Gordon distance}
\usage{
dissgordon(X)
}
\arguments{
\item{X}{Numerical matrix with integer/ordinal values.}
}
\value{
Distance matrix
}
\description{
Gordon distance for ordinal data
}
