% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{disshamming}
\alias{disshamming}
\title{Hamming distance wrapper function}
\usage{
disshamming(X)
}
\arguments{
\item{X}{matrix}
}
\value{
Distance matrix
}
\description{
Function based on cultevo's package implementation
}
