% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{disshellinger}
\alias{disshellinger}
\title{Hellinger distance (wrapper)}
\usage{
disshellinger(X)
}
\arguments{
\item{X}{Matrix}
}
\value{
Distance R object
}
\description{
Wrapper of `HellingerDist`
}
