% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{dissks}
\alias{dissks}
\title{KS ordinal pairwise distance}
\usage{
dissks(X)
}
\arguments{
\item{X}{Numerical matrix with integer/ordinal values.}
}
\value{
Distance matrix
}
\description{
KS-ordinal pairwise function
}
