% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{disspodani}
\alias{disspodani}
\title{Podani's distance function}
\usage{
disspodani(X, M = NULL)
}
\arguments{
\item{X}{Numerical matrix with integer/ordinal values.}

\item{M}{numeric Number of largest category. If `M = NULL`, `M` is determined
for each variable of the data.}
}
\value{
Distance matrix
}
\description{
R implementation of Podani's distance.
}
