% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{dissspearman}
\alias{dissspearman}
\title{Spearman distance wrapper function}
\usage{
dissspearman(X)
}
\arguments{
\item{X}{matrix}
}
\value{
Distance matrix
}
\description{
Function based on bioDist's package implementation.
}
