% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{disstau}
\alias{disstau}
\title{Kendall's Tau distance wrapper function}
\usage{
disstau(X)
}
\arguments{
\item{X}{Numerical matrix with integer/ordinal values.}
}
\value{
Distance matrix
}
\description{
Function based on bioDist's package implementation.
}
