% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{dissWass}
\alias{dissWass}
\title{Wasserstein's distance function}
\usage{
dissWass(X, type = "DS", m = NULL, p = 1)
}
\arguments{
\item{X}{numerical matrix containing integer/ordinal values.}

\item{type}{character Select between three WD implementations: \code{KS} (Known Support),
\code{DS} (Data Support), \code{R} (Rank).}

\item{m}{integer Specifies the number of nominal categories for \code{type='KS'} option.}

\item{p}{numeric Order of the distance. p=1 by default}
}
\value{
Distance matrix
}
\description{
Wrapper for the proposed Wasserstein's distance types
}
