% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{dissWassDS}
\alias{dissWassDS}
\title{Wasserstein's DS distance function}
\usage{
dissWassDS(X, p = 1)
}
\arguments{
\item{X}{numeric Matrix with integer values.}

\item{p}{numeric Order of the distance. p=1 by default.}
}
\value{
Distance matrix
}
\description{
Estimates eCDF based on the union of observed values
}
