% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{dissWassKS}
\alias{dissWassKS}
\title{Wasserstein's KS distance function}
\usage{
dissWassKS(X, maxM, p = 1)
}
\arguments{
\item{X}{numeric Matrix with integer values.}

\item{maxM}{integer Number of ordinal values which will range from 1 to \code{maxM}}

\item{p}{numeric Order of the distance. p=1 by default}
}
\value{
Distance matrix
}
\description{
Assumes known and identical number of ordinal categories \code{maxM}
}
