% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{distancematrix}
\alias{distancematrix}
\title{Calculate categorical distance matrix for discrete data}
\usage{
distancematrix(X, d)
}
\arguments{
\item{X}{Matrix where rows are observations and columns are discrete features.}

\item{d}{Character scalar naming the distance.}
}
\value{
An object of class \code{dist}.
}
\description{
Dispatcher for discrete distance functions (nominal + ordinal).
}
\examples{
X <- rbind(matrix(paste0("a", rpois(7*5, 1)), nrow=5),
           matrix(paste0("a", rpois(7*5, 3)), nrow=5))
distancematrix(X = X, d = "hellinger")
}
