% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteClusGapFuns.R
\name{findK}
\alias{findK}
\title{Criteria to determine number of clusters k}
\usage{
findK(cG_obj, meth = "Tibs2001SEmax", SE.fact = 1)
}
\arguments{
\item{cG_obj}{Output object obtained from \code{clusGapDiscr}}

\item{meth}{Method to use to determine optimal k number of clusters.}

\item{SE.fact}{Standard Error Factor generalizing the 1-SE rule.}
}
\value{
A numerical value from 1 to K.max, contained in the input \code{cG_obj} object.
}
\description{
Same function as found in \code{cluster} package.
}
