% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFuns.R
\name{kmodesD}
\alias{kmodesD}
\title{Adapted k-modes algorithm}
\usage{
kmodesD(data, modes, distFun, iter.max = 10)
}
\arguments{
\item{data}{A matrix or data frame of categorical data. Objects have to be in rows, variables in columns.}

\item{modes}{The number of modes}

\item{distFun}{Pairwise categorical distance function. A function accepting two categorical vectors.}

\item{iter.max}{The maximum number of iterations allowed.}
}
\value{
An object of class kmodes as found in `klaR` packages.
An additional component specifies the categorical distance function found in `distFun`.
}
\description{
K-modes function to accept any categorical distance based on
the function found in `klaR:kmodes`.
}
