% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mass}
\alias{mass}
\title{mass data}
\format{
## `data.frame`
\describe{
  \item{Gender}{Gender}
  \item{I find math interesting.}{Math interesting}
  \item{I get uptight during math tests.}{Uptight with math tests}
  \item{I think that I will use math in the future.}{Use math in the future}
  \item{Mind goes blank and I am unable to think clearly when doing my math test.}{Mind goes blank in math tests}
  \item{Math relates to my life.}{Math relates to own life}
  \item{I worry about my ability to solve math problems.}{Worry about ability math problem solving}
  \item{I get a sinking feeling when I try to do math problems.}{Sinking feeling doing math problems}
  \item{I find math challenging.}{Math is challenging}
  \item{Mathematics makes me feel nervous.}{Nervousness with math}
  \item{I would like to take more math classes.}{Take more math classes}
  \item{Mathematics makes me feel uneasy.}{Uneasy feeling with math}
  \item{Math is one of my favorite subjects.}{Favorite subject is math}
  \item{I enjoy learning with mathematics.}{Enjoy learning math}
  \item{Mathematics makes me feel confused.}{Confused with math}
}
}
\source{
<https://rdrr.io/cran/likert/man/mass.html>
}
\usage{
mass
}
\description{
Data extracted from the `likert` R package.
Results from an administration of the Math Anxiety Scale Survey.
First Column records student gender either Female or Male.
All statement answers have 5 possible ordinal categorical items:
Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree.
}
\keyword{datasets}
