% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFuns.R
\name{plotMDS2}
\alias{plotMDS2}
\title{MDS Plots for Categorical Data}
\usage{
plotMDS2(
  x,
  cl,
  type = "MDS",
  cols = NULL,
  dotSize = 3,
  LabTitle = type,
  outDir = "./",
  filename = paste0(type, "plot.png"),
  addRowNames = FALSE,
  labSize = 3,
  out = "plot"
)
}
\arguments{
\item{x}{dist object}

\item{cl}{character Vector with clustering assignments. Important: the vector order should match
number (and if possible names) as the observed data matrix used to generate the distance object `x`.}

\item{type}{character String specifying the class of MDS: either classic `'MDS'` (default) or
Non-metrical Dimensional Scaling `'NMDS'`. The core `MDS` function used is `stats::cmdscale(x, k = 2)` and
`vegan::metaMDS(comm = x, distance = "none", k = 2, trymax = 100, autotransform = FALSE)` for `NMDS`,
which uses `vegan::monoMDS` by default.}

\item{cols}{character Vector of colors to use with clustering labels. Cluster names should match.
to the ones provided in `cl`. The by default (`cols = NULL`) de functions produces highly contrasting colors.}

\item{dotSize}{numerical String specifying the point size. `dotSize = 3` by default.}

\item{LabTitle}{character String for the plot's title. By default `LabTitle = type`.}

\item{outDir}{character string with the directory path to save output file. `outDir = './'` by default.}

\item{filename}{character string with name of file output. `filename = paste0(type, 'plot.png')` by default.}

\item{addRowNames}{logical Single value indicating whether to place observation names next to points.
The labels used are the names found in `cl`. If `names(cl) == NULL`, the samples
will be labelled in number of appearance. ggrepel package is used to locate the labels.}

\item{labSize}{numeric Single value indicating the size of labels. `labSize = 3` by default.
Non-functional if `addRowNames = FALSE`.}

\item{out}{character String specify output to obtain: `'plot'` (default), `'data.frame'` or `'object'`. Either
`plot` return the resulting ggplot object or
`data.frame` for a data.frame with coordinates and corresponding clustering assignments or
`object` to obtain raw objects from the chosen `type` above.}
}
\value{
png file and either a data.frame with coordinates and labels or either a list with
related MDS results or a `metaMDS` object, depending on the `out` option.
}
\description{
Function to visualize the distribution and spread of categorical data in two dimensions
starting from the distance matrix.
A cluster assignment vector needs to be provided to show color coded points.
}
