% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteClusGapFuns.R
\name{clusGapUDiscr}
\alias{clusGapUDiscr}
\title{Discrete application of clusGap
Based on the implementation of the function found in the `cluster` R package}
\usage{
clusGapUDiscr(
  x,
  FUNcluster,
  K.max,
  B = 100,
  vals = NULL,
  verbose = interactive(),
  distName = "hamming",
  ...
)
}
\arguments{
\item{x}{categorical/number matrix}

\item{FUNcluster}{a function that accepts as first argument a matrix like `x`; second argument specifies number of `k` (k=>2) clusters
This functions returns a list with a component named `cluster`, a vector of length `n=nrow(x)` of integers from `1:k` indicating observation cluster assignment.}

\item{K.max}{Maximum number of clusters `k` to consider}

\item{B}{Number of bootstrap samples.}

\item{vals}{String or numerical vector. A vector with all categories to consider when bootstrapping the null distribution sample (FR: Full Range option).
By DEFAULT vals=NULL, meaning unique range of categories found in the data will be used when drawing the null (DR: Data Range option).}

\item{verbose}{integer or logical determining whether progress output should printed while running. By DEFAULT one bit is printed per bootstrap sample.}

\item{distName}{Name of categorical distance to apply}

\item{...}{optionally further arguments for `FUNcluster()`}

\item{integer}{or logical, determining if “progress” output should be printed. The default prints one bit per bootstrap sample.}
}
\value{
a matrix with K.max rows and 4 columns, named "logW", "E.logW", "gap", and "SE.sim",
where gap = E.logW - logW, and SE.sim correspond to the standard error of `gap`.
}
\description{
Discrete application of clusGap
Based on the implementation of the function found in the `cluster` R package
}
