% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFuns.R
\name{clusterFunSel}
\alias{clusterFunSel}
\title{Clustering generating function}
\usage{
clusterFunSel(clustFun)
}
\arguments{
\item{clustFun}{A character string with the following possible options:
'pam' (default) from `cluster::pam`, 'diana' from `cluster::diana`, 'fanny' from `cluster::fanny`,
'agnes-\{average, single, complete, ward, weighted\}' from `cluster::agnes`,
'hclust-\{ward.D, ward.D2, single, complete, average, mcquitty, median, centroid\}' from `base::hclust`,
'kmodes' from `klar::kmodes` (`iter.max = 10`, `weighted = FALSE` and `fast= TRUE`).
'kmodes-N' enables to run the `kmodes` algorithm with a given number N of iterations where `iter.max = N`.}
}
\value{
An object of class kmodes as found in `klaR` packages.
An additional component specifies the categorical distance function found in `distFun`.
}
\description{
A function that generates formatted algorithmic functions that can be plugged
to enable run a wide variety of clustering algorithm for `clusGapDiscr` function.
}
