% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{dissbhattacharyya}
\alias{dissbhattacharyya}
\title{Bhattacharyya's wrapper Function}
\usage{
dissbhattacharyya(X, na.rm = TRUE)
}
\arguments{
\item{X}{Matrix}

\item{na.rm}{Remove NAs default=TRUE}

\item{adj}{Small quantity added to avoid indefinite log(0) values. DEFAULT=0.001}
}
\value{
Distance R object
}
\description{
Bhattacharyya's wrapper Function
}
