% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{disshamming}
\alias{disshamming}
\title{Hamming distance wrapper function
Function based on cultevo's package implementation}
\usage{
disshamming(X, na.rm = TRUE)
}
\arguments{
\item{X}{matrix}

\item{na.rm}{logical}
}
\value{
Distance matrix
}
\description{
Hamming distance wrapper function
Function based on cultevo's package implementation
}
