% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{disshellinger}
\alias{disshellinger}
\title{Hellinger's distance wrapper Function}
\usage{
disshellinger(X, na.rm = TRUE)
}
\arguments{
\item{X}{Matrix}

\item{na.rm}{logical}
}
\value{
Distance R object
}
\description{
Hellinger's distance wrapper Function
}
