% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{disspodani}
\alias{disspodani}
\title{####################
Ordinal distances ##
####################
Podani's distance function}
\usage{
disspodani(X, M = NULL)
}
\arguments{
\item{X}{matrix}
}
\value{
Distance matrix
}
\description{
R implementation of Podani's distance.
}
