% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{disspodani}
\alias{disspodani}
\title{Podani distance wrapper function}
\usage{
disspodani(X)
}
\arguments{
\item{X}{Numerical matrix with integer/ordinal values.}
}
\value{
Distance matrix
}
\description{
Function based on FD's implementation
}
