% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFuns.R
\name{dissWass}
\alias{dissWass}
\title{Wasserstein's distance function}
\usage{
dissWass(X, p = 1)
}
\arguments{
\item{X}{Numerical matrix with integer/ordinal values.}

\item{p}{numeric Order of the distance}
}
\value{
Distance matrix
}
\description{
Using transport's R package implementation
}
