% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteClusGapFuns.R
\name{findK}
\alias{findK}
\title{Criteria to determine number of clusters k}
\usage{
findK(cG_obj, meth = "Tibs2001SEmax")
}
\arguments{
\item{cG_obj}{Output object obtained from `clusGapDiscr`}

\item{meth}{Method to use to determine optimal k number of clusters.}
}
\description{
Criteria to determine number of clusters k
}
