% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFuns.R
\name{plotMDS2}
\alias{plotMDS2}
\title{MDS Plots for Categorical Data}
\usage{
plotMDS2(
  x,
  cl,
  type = "MDS",
  cols = NULL,
  dotSize = 3,
  LabTitle = type,
  outDir = NULL,
  filename = NULL,
  addRowNames = FALSE,
  labSize = 3,
  out = "plot"
)
}
\arguments{
\item{x}{dist object}

\item{cl}{character Vector with clustering assignments. Important: the vector order should match
number (and if possible names) as the observed data matrix used to generate the distance object \code{x}.}

\item{type}{character String specifying the class of MDS: either classic \code{'MDS'} (default) or
Non-metrical Dimensional Scaling \code{'NMDS'}. The core \code{MDS} function used is \code{stats::cmdscale(x, k = 2)} and
\code{vegan::metaMDS(comm = x, distance = "none", k = 2, trymax = 100, autotransform = FALSE)} for \code{NMDS},
which uses \code{vegan::monoMDS} by default.}

\item{cols}{character Vector of colors to use with clustering labels. Cluster names should match.
to the ones provided in \code{cl}. The by default (\code{cols = NULL}) de functions produces highly contrasting colors.}

\item{dotSize}{numerical String specifying the point size. \code{dotSize = 3} by default.}

\item{LabTitle}{character String for the plot's title. By default \code{LabTitle = type}.}

\item{outDir}{character string with the directory path to save output file.
\code{outDir = NULL} is default option avoiding generating a plot.}

\item{filename}{character string with name of file output.
\code{filename = NULL} by default avoiding generating a plot.
This parameter should have a valid \code{ggplot2} output graphical format extension like \code{'FileName.{png,pdf,ps,...}'}.}

\item{addRowNames}{logical Single value indicating whether to place observation names next to points.
The labels used are the names found in \code{cl}. If \code{names(cl) == NULL}, the samples
will be labelled in number of appearance. ggrepel package is used to locate the labels.}

\item{labSize}{numeric Single value indicating the size of labels. \code{labSize = 3} by default.
Non-functional if \code{addRowNames = FALSE}.}

\item{out}{character String specify output to obtain: \code{'plot'} (default), \code{'data.frame'} or \code{'object'}. Either
\code{plot} return the resulting ggplot object or
\code{data.frame} for a data.frame with coordinates and corresponding clustering assignments or
\code{object} to obtain raw objects from the chosen \code{type} above.}
}
\value{
png file and either a data.frame with coordinates and labels or either a list with
related MDS results or a \code{metaMDS} object, depending on the \code{out} option.
}
\description{
Function to visualize the distribution and spread of categorical data in two dimensions
starting from the distance matrix.
A cluster assignment vector needs to be provided to show color coded points.
}
