% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFuns.R
\name{ResHeatmap}
\alias{ResHeatmap}
\title{Heatmap assuming a given a distance function and a known number of clusters.
Function to display a categorical data matrix given a user defined number of clusters `nCl`, a categorical distance `distName` and a predefined clustering method `FUNcluster`.
The output displays a heatmap separating and color-labelling resulting clusters vertically in the rows and allowing unsupervised clustering on questions in the columns. Each cell is colored according to the categorical values provided or found in the data.
The clustergram is based on the `pheatmap` function from the pheatmap R package. Thus, any parameter found in pheatmap can be specified to `clusGapDiscrHeat`.
This function can be used to examine number of clusters before running `clusGapDiscrHeat` but also after number of clusters is determined.}
\usage{
ResHeatmap(
  x,
  nCl,
  distName,
  catVals,
  FUNcluster = cluster::pam,
  out = "heatmap",
  clusterNames = NULL,
  prefObs = NULL,
  filename = NULL,
  outDir = NULL,
  height = 10,
  width = 6
)
}
\arguments{
\item{x}{matrix object or data.frame}

\item{nCl}{number of clusters to plot; if `nCl` is a vector, its length lN is the number of clusters, with values ranging from 1 to lN in desired order to place the clusters.}

\item{distName}{Name of categorical distance to apply.
Available distances: 'bhattacharyya', 'chisquare', 'cramerV', 'hamming' and 'hellinger'.}

\item{catVals}{character string vector with (ordered) categorical values}

\item{FUNcluster}{a function that accepts as first argument a matrix like `x`; second argument specifies number of `k` (k=>2) clusters
This function returns a list with a component named `cluster`, a vector of length `n=nrow(x)` of integers from `1:k` indicating observation cluster assignment.}

\item{out}{Specifies the desired output between "heatmap" (default; produce a heatmap), "clusters" (return a `data.frame` with clustering assignments) or "clustersReord" (return a `data.frame` with reorganized clusters)}

\item{clusterNames}{Either `null` or 'renumber'. When `nCl` the cluster ordering is rearranged. `NULL` leaves cluster names as their original cluster assignment. 'renumber' respects the rearrangements but relabels the cluster numbers from top to bottom in ascending order.}

\item{prefObs}{character string vector of length 1 with a prefix for the observations, in case they come unlabelled or the user wants to anomymize sample IDs}

\item{filename}{character string with name of file output}

\item{outDir}{character string with the directory path to save output file}

\item{height}{numeric width of output plot in inches}
}
\value{
png file or ComplexHeatmap object
}
\description{
Heatmap assuming a given a distance function and a known number of clusters.
Function to display a categorical data matrix given a user defined number of clusters `nCl`, a categorical distance `distName` and a predefined clustering method `FUNcluster`.
The output displays a heatmap separating and color-labelling resulting clusters vertically in the rows and allowing unsupervised clustering on questions in the columns. Each cell is colored according to the categorical values provided or found in the data.
The clustergram is based on the `pheatmap` function from the pheatmap R package. Thus, any parameter found in pheatmap can be specified to `clusGapDiscrHeat`.
This function can be used to examine number of clusters before running `clusGapDiscrHeat` but also after number of clusters is determined.
}
