% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDataFun.R
\name{SimData}
\alias{SimData}
\title{Simulate Data}
\usage{
SimData(N, nQ, pi)
}
\arguments{
\item{N}{Integer. Number of observations.}

\item{nQ}{Integer. Number of questions.}

\item{pi}{Numeric vector. Vector of probabilities adding up to 1. Names of elements can be characters.}
}
\value{
N x nQ matrix with simulated categories distributed according to vector pi
}
\description{
Simulate Data
}
\examples{
Pix <- setNames(c(0.1, 0.2, 0.3, 0.4, 0), paste0('a', 1:5))
X <- SimData(N=10, nQ=5, Pix)
head(X)

Piy <- setNames(c(0.6, 0.1, 0.1, 0.1, 0.1), letters[1:5])
Y <- SimData(N=10, nQ=3, Piy)
head(Y)
}
